engine/requirements.txt
requests>=2.31.0
praw>=7.7.0
pytrends>=4.9.0
python-dotenv>=1.0.0
supabase>=2.0.0
google-api-python-client>=2.0.0
